# triton-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Triton Backend",
    "description": "Backend for Triton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/instances": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Create Instance",
        "description": "Create a Triton instance from UI input (url + optional name).\n\nThe backend connects to Triton, reads the currently available models,\nstores the instance in the DB, and returns the created DTO.",
        "operationId": "create_instance_api_instances_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTritonInstanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "List Instances",
        "description": "List all registered Triton instances.",
        "operationId": "list_instances_api_instances_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TritonInstanceDTO"
                  },
                  "title": "Response List Instances Api Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances/{instance_id}/models": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance Models",
        "description": "Return the live Triton model repository index for an instance.",
        "operationId": "get_instance_models_api_instances__instance_id__models_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TritonRepositoryModelDTO"
                  },
                  "title": "Response Get Instance Models Api Instances  Instance Id  Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/models/{model_name}/versions/{version}/config": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance Model Config",
        "description": "Proxy the live Triton model config for a specific model version.",
        "operationId": "get_instance_model_config_api_instances__instance_id__models__model_name__versions__version__config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Model Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Instance Model Config Api Instances  Instance Id  Models  Model Name  Versions  Version  Config Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/models/{model_name}/load": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Load Instance Model",
        "description": "Trigger explicit load for a model via Triton's repository API.",
        "operationId": "load_instance_model_api_instances__instance_id__models__model_name__load_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRepositoryActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/models/{model_name}/unload": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Unload Instance Model",
        "description": "Trigger explicit unload for a model via Triton's repository API.",
        "operationId": "unload_instance_model_api_instances__instance_id__models__model_name__unload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRepositoryActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/models/{model_name}/versions/{version}/infer": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Infer Instance Model",
        "description": "Proxy a Triton inference request for a specific model version.",
        "operationId": "infer_instance_model_api_instances__instance_id__models__model_name__versions__version__infer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Model Name"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload",
                "description": "Triton inference request body."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTritonInstanceRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": false
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreateTritonInstanceRequest",
        "description": "Request to create a Triton instance."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstanceS3ConfigDTO": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix"
          },
          "access_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key"
          },
          "secret_configured": {
            "type": "boolean",
            "title": "Secret Configured",
            "default": false
          },
          "use_https": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Https"
          },
          "verify_ssl": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Ssl"
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "address_style": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "path",
                  "virtual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Style"
          }
        },
        "type": "object",
        "title": "InstanceS3ConfigDTO"
      },
...
      "ModelRepositoryActionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ModelRepositoryActionResponse"
      },
...
      "TritonInstanceDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "model_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Names"
          },
          "repository_models": {
            "items": {
              "$ref": "#/components/schemas/TritonRepositoryModelDTO"
            },
            "type": "array",
            "title": "Repository Models"
          },
          "server_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Metadata"
          },
          "health_live": {
            "type": "boolean",
            "title": "Health Live",
            "default": false
          },
          "health_ready": {
            "type": "boolean",
            "title": "Health Ready",
            "default": false
          },
          "health_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Last Checked At"
          },
          "health_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Error"
          },
          "triton_verify_ssl": {
            "type": "boolean",
            "title": "Triton Verify Ssl",
            "default": false
          },
          "triton_ca_certificate": {
            "type": "string",
            "title": "Triton Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          },
          "metrics_cpu": {
            "type": "number",
            "title": "Metrics Cpu",
            "default": 0
          },
          "metrics_ram": {
            "type": "number",
            "title": "Metrics Ram",
            "default": 0
          },
          "metrics_gpu": {
            "type": "number",
            "title": "Metrics Gpu",
            "default": 0
          },
          "metrics_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Last Checked At"
          },
          "metrics_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Error"
          },
          "deployment_runtime": {
            "type": "string",
            "title": "Deployment Runtime",
            "default": "external"
          },
          "deployment_namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Namespace"
          },
          "deployment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Name"
          },
          "deployment_service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Service Name"
          },
          "deployment_secret_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Secret Name"
          },
          "deployment_log": {
            "type": "string",
            "title": "Deployment Log",
            "default": ""
          },
          "is_self_deployed": {
            "type": "boolean",
            "title": "Is Self Deployed",
            "default": false
          },
          "pod_statuses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pod Statuses"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "s3": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstanceS3ConfigDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "name",
          "model_names",
          "created_at"
        ],
        "title": "TritonInstanceDTO",
        "description": "DTO for outgoing Triton instance data (API responses)."
      },
      "TritonRepositoryModelDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TritonRepositoryModelDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}